chore: move navigation drawer to admin ui
