+++ apps/journeys-admin/src/components/PageWrapper/PageWrapper.tsx
@@ -1,6 +1,7 @@
 import Box from '@mui/material/Box'
 import Stack from '@mui/material/Stack'
 import { useTheme } from '@mui/material/styles'
+import { NavigationDrawer } from '@core/admin/ui/NavigationDrawer'
 import { useRouter } from 'next/router'
 import { User } from 'next-firebase-auth'
 import { ReactElement, ReactNode, useState } from 'react'
@@ -11,7 +12,6 @@ import { PageProvider, PageState } from '../../libs/PageWrapperProvider'
 import { AppHeader } from './AppHeader'
 import { MainPanelBody } from './MainPanelBody'
 import { MainPanelHeader } from './MainPanelHeader'
-import { NavigationDrawer } from './NavigationDrawer'
 import { SidePanel } from './SidePanel'
 import { usePageWrapperStyles } from './utils/usePageWrapperStyles'
