+++ libs/admin/ui/i18next-parser.config.js
@@ -2,6 +2,6 @@ const i18nextParserConfigBase = require('../../../i18next-parser.config.base')
 
 module.exports = {
   ...i18nextParserConfigBase,
-  defaultNamespace: 'libs-journeys-admin-ui',
+  defaultNamespace: 'libs-admin-ui',
   input: ['src/**/*.{js,jsx,ts,tsx}']
 }
