+++ libs/admin/ui/src/components/NavigationDrawer/UserNavigation/ImpersonateDialog/ImpersonateDialog.stories.tsx
@@ -2,7 +2,7 @@ import { Meta, StoryObj } from '@storybook/react'
 import { screen, userEvent } from '@storybook/testing-library'
 import { ReactElement, useState } from 'react'
 
-import { journeysAdminConfig } from '../../../../../libs/storybook'
+import { journeysAdminConfig } from '../../../../../../../../apps/journeys-admin/src/libs/storybook'
 
 import { ImpersonateDialog, USER_IMPERSONATE } from './ImpersonateDialog'
