+++ libs/admin/ui/src/components/NavigationDrawer/UserNavigation/UserMenu/UserMenu.stories.tsx
@@ -2,7 +2,7 @@ import { Meta, StoryObj } from '@storybook/react'
 import noop from 'lodash/noop'
 import { ReactElement, useState } from 'react'
 
-import { simpleComponentConfig } from '../../../../../libs/storybook'
+import { simpleComponentConfig } from '../../../../../../../../apps/journeys-admin/src/libs/storybook'
 
 import { UserMenu } from '.'
