+++ libs/admin/ui/src/components/NavigationDrawer/UserNavigation/UserMenu/index.ts
