--- libs/journeys-admin/ui/README.md
@@ -1,11 +1,0 @@
-# journeys-admin-ui
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test journeys-admin-ui` to execute the unit tests via [Jest](https://jestjs.io).
-
-## Running storybook
-
-Run `nx demo journeys-admin-ui` for storybook server. Navigate to http://localhost:4400/. The app will automatically reload if you change any of the source files.
